Sorting Long arrays instead of Integer arrays

+++ src/sorting/SelectionSort.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class SelectionSort implements ISortAlgorithm{
 
     @Override
-    public List<Integer> sort(List<Integer> array) {
+    public List<Long> sort(List<Long> array) {
         int n = array.size();
         for (int i = 0; i < n - 1; i++) {
             int minIndex = i;
@@ -14,7 +14,7 @@ public class SelectionSort implements ISortAlgorithm{
                     minIndex = j;
                 }
             }
-            int temp = array.get(minIndex);
+            Long temp = array.get(minIndex);
             array.set(minIndex, array.get(i));
             array.set(i, temp);
         }
